Move fields into /api/v1/configs/fields
